Fix signed/unsigned
Former-commit-id: 198de3d148b08078fef8a95dc781ba7cbd111461 [formerly 80c8beb86d35c68c44d73e14ae0064129a123571]
Former-commit-id: 44d6d3cd63464809c2790688c4495a1cc1fdcbef

+++ apps/linear_blur/simple_blur_generator.cpp
@@ -27,14 +27,14 @@ public:
             // Might be x,y but might be _0,_1. Use the args() to work around.
             input.estimate(input.args()[0], 0, W)
                  .estimate(input.args()[1], 0, H);
-            for (int i = 2; i < input.args().size(); ++i) {
+            for (size_t i = 2; i < input.args().size(); ++i) {
                 input.estimate(input.args()[i], 0, C);
             }
             width.set_estimate(W);
             height.set_estimate(H);
             output.estimate(x, 0, W)
                   .estimate(y, 0, H);
-            for (int i = 2; i < output.args().size(); ++i) {
+            for (size_t i = 2; i < output.args().size(); ++i) {
                 output.estimate(output.args()[i], 0, C);
             }
         } else {
